feat: remove with_base() method to simplify API
BREAKING CHANGE: Removed AppPath::with_base() static method

- Simplifies API by focusing on core use case: paths relative to executable
- Users can use standard Path::join() for custom base directories
- Reduces API surface area and cognitive load
- Updated all documentation and examples
- Updated all tests to use standard path joining patterns

For custom base directories, users should now use:
  let custom_path = base_dir.join('config.toml');
  let app_path = AppPath::new(custom_path);

This change makes the API more focused and eliminates confusion about
when to use new() vs with_base().
